improve the fetch_versions client (#114)
Update the fetch_versions client to support fetching from the top-level
directory (so contrib versions can be updated too). Add support for
using GITHUB_TOKEN for authentication (to avoid rate limit issues). Do a
replace across all files of the old version with the new version when
creating the files for a new plugin.