update fetcher to retrieve latest version (#145)
Maven versions can be returned out of order (the default sort is by
score). Update the query to determine the latest semver version from the
API (ignoring pre-release versions).

Additionally, fetch the latest patch version for each major+minor (we
can update the utility in the future to use this to update to later
patch versions of plugins we already ship).