configure go module cache before installing buf CLI (#178)
In order to use newer features of the buf CLI that haven't been released
yet, we're building the CLI from the main branch. However, we're
restoring Go module cache after we install the CLI which means that
we're not benefiting from previous builds.

Update the workflows to configure the Go module cache before running 'go
install'.