automate more steps in fetcher command
Update fetcher to run 'go mod tidy', 'npm install', and run the tests to
generate a plugin.sum file (and show the differences to generated code).
This should automate almost all steps when updating to a new version of
a plugin except for updating Docker base images or dependencies on other
plugins/runtime libraries.