create Go CLI to automate plugin releases
Create a Go CLI (cmd/release/main.go) which automates the creation of
plugin releases. A release consists of the following:

* plugin-releases.json - Contains metadata about all plugins (including
  name, version, digests).
* plugin-releases.json.minisign - Minisign signature of
  plugin-releases.json.
* <owner>-<name>-<version>.zip - One or more zip files containing a
  buf.plugin.yaml and image.tar (created with docker save).

The CLI is configured to run once daily. It detects any changes since
the latest release and will upload only the added/changed plugins.

There are two supported options to the CLI:

* `-minisign-private-key <key>`: Path to minisign private key. If
  omitted, the plugin-releases.json file will not be signed.
* `-dry-run`: If true, the release will be created locally and saved to
  a temporary directory on disk, but no changes will be made to GitHub.