switch to smaller base images (#228)
As a prereq to packaging plugins for installation, we should switch to
smaller base images. This will also improve security by not including a
full OS in our plugins.

Overall savings varies by plugins with Java and Node plugins being the
largest. In the best case, we see ~93% size reduction in some images.