Add restore-release utility
Add a restore-release utility which can automate restoring the container
registry to the same images and tags as a previous release. This can be
used to avoid unnecessary plugin installations and revision bumps for
plugins whose Dockerfile or metadata didn't change since the last
release (in this case, due to broken automation with a
tj-actions/changed-files action).