Simplify plugins build process (#453)
Add support for workflow_dispatch (manual builds) to specify the PLUGINS
env var to allow for rebuilding certain plugins (bypassing the default
logic that uses changed-files actions to determine what to rebuild).

Update the changed-files action to run just once and print out a value
which can be used as the PLUGINS env var. This allows us to isolate this
logic around github actions to just a single place in the code.

Update places that decide which plugins to run to be more conservative.
If no filter parameters are used, don't build any plugins and show an
error message. This should make it better for new contributors to this
project to more easily get started with building a specific plugin
(since no one should really be building all 300+ plugins at this point
unless they really mean to do so).