Add kotlin-stdlib in addition to jdk8 variant v3
We should ensure that we pull in both kotlin-stdlib AND
kotlin-stdlib-jdk8 to ensure that consistent versions are used with the
dependencies and the compiler. Without this, it is possible that
protobuf-kotlin's dependency version kotlin-stdlib 1.6.0 is used
instead.

Breaking this change into multiple PRs as these plugins take a long time to build.