Only rebuild plugins if changed
Update the dockerbuild command to add a label containing the git
revision of the plugin's directory (if there are no changes to the
working directory of the plugin). If we determine that an image exists
for the same revision of the plugin, skip costly rebuilds.

Additionally, update commands to use exec.CommandContext consistently
and stop calling exec.LookPath as the exec package does this
automatically for us if the command doesn't contain path separators.