Add connectrpc/swift plugin dependency for connectrpc/swift-mocks
This plugin depends on the output of the connect-swift generated code,
so it needs a plugin dependency here to behave correctly with generated
SDKs.

This is only currently fixing this issue in the latest connect-swift
release; not sure if we want to backport it (I'm not clear on if this
has always been an issue or was introduced in a particular update, but
would lean towards the former?).

Ref: https://github.com/connectrpc/connect-swift/issues/250