Add pb/py<->pb/pyi dependencies
Discussed this internally - we made the decision when releasing the
Python Repository to not add a dependency between these plugins,
specifically in the python->pyi direction, so that users that don't want
type annotations don't need to have them included as a
dependency.

However, if you _want_ type annotations, it's confusing to need to
install both packages. Also, a `*_pyi` package is useless on its own
(since it creates a [stub-only package][1]). Lastly, grpc/python comes
with stubs built-in now, and protocolbuffers leans pretty heavily into
types. Let's add this dependency relationship moving forward to simplify
things.

For now, only adding this on the latest version - we can backport if
there's significant interest.

[1]: https://peps.python.org/pep-0561/#stub-only-packages