Require protobuf dep to be at least gencode level
Update the python plugins to ensure we require a dependency version of
the protobuf library to be at least the same version of the gencode.
Without this constraint, users that have an earlier 5.27.0 or 5.27.1
version of the protobuf dependency already installed will fail at
runtime.

See https://protobuf.dev/support/cross-version-runtime-guarantee/ for
more details.