Update latest-plugins to support include file
Update the latest-plugins command to support an --include-file
parameter, which when used will include all of the plugin names/versions
from the file in addition to the latest versions of all plugins in the
output.

This also ensures that if there are newer revisions to existing plugins,
they'll also be updated to the latest version.