Ensure bazel server starts on IPv4 only
When building on macOS with the docker-container driver, there are
errors connecting between the server and client used internally by
bazel. It appears that the server binds to '[::1]:0' and the client for
some reason can't connect.

Update to specify that the server should prefer IPv4 over IPv6
addresses, which appears to resolve the issue.