Disable grpc/swift fetcher
With v2.0.0, the protoc plugin for grpc-swift now exists in a separate
grpc-swift-protobuf project. Unfortunately, the two projects are no
longer versioned together, so we'll need to brainstorm the best way
to support this plugin going forward.

For now, disable updates for this plugin to allow other updates to
proceed.