Change Python version specifiers for protobuf to `>=`
The previous specifiers were too restrictive of maximum values; we
_know_ the minimum version that we want to allow, but we leave it up to
users to determine their maximum. (Newer versions of the `protobuf`
runtime should support previous versions of gencode, [at least up to the
next major version][1], but possibly further in practice.)

In terms of `types-protobuf`, the versions simply match the version of
the runtime, and so similarly should not have an upper bound.

[1]: https://protobuf.dev/support/cross-version-runtime-guarantee/#major