Loosen protobuf dependency to >3.20
We currently set a minimum version of the runtime `protobuf` package for
protocolbuffers/python based on the documented ["new gencode + old
runtime = never allowed"][1] rule that generally applies to the
protocolbuffers plugins - that the runtime versions are free to add
runtime APIs in any release and it's expected that the gencode must be
updated to run with the new runtime version.

However, [Python has a separate carve-out behavior for it's gencode][2],
documented now as:

> Since the 3.20.0 release, the Protobuf Python generated code became a
> thin wrapper around an embedded FileDescriptorProto. Because these
> protos are supported on extremely long timeframes, our usual major
> version compatibility windows aren’t typically necessary.

> Python may break generated code compatibility in specific future major
> version releases, but it will be coupled with poison pill warnings and
> errors in advance. As of 6.32.0, all generated code since 3.20.0 will be
> supported until at least 8.x.y.

This means that we don't need to pin the protocolbuffers/python version
to the matching runtime version; we can loosen the minimum to `3.20.0`
(and will need to reevaluate this in the future v8 major release).

This commit pins the latest minor release in each major version to a
minimum of 3.20.

An alternative is to _just_ specify `protobuf` without any version
specifiers, and allow consumers to pick the version of `protobuf` they
want to supply. But since the minimum is pretty clearly documented, I
think supplying it is reasonable.

[1]: https://protobuf.dev/support/cross-version-runtime-guarantee/#backwards
[2]: https://protobuf.dev/support/cross-version-runtime-guarantee/#python